+++ src/App.vue
@@ -1,18 +1,9 @@
 <template>
-  <div
-    v-if="isAppInitialized"
-    class="app__container"
-  >
+  <div v-if="isAppInitialized" class="app__container">
     <app-navbar class="app__navbar" />
     <router-view v-slot="{ Component, route }">
-      <transition
-        :name="route.meta.transition || 'fade'"
-        mode="out-in"
-      >
-        <component
-          class="app__main"
-          :is="Component"
-        />
+      <transition :name="route.meta.transition || 'fade'" mode="out-in">
+        <component class="app__main" :is="Component" />
       </transition>
     </router-view>
   </div>
@@ -28,9 +19,8 @@ import { useNotifications } from '@/composables'
 export default defineComponent({
   name: 'app',
   components: { AppNavbar },
-  setup () {
+  setup() {
     const isAppInitialized = ref(false)
-
     const init = async () => {
       try {
         useNotifications()