+++ src/validators.ts
@@ -1,5 +1,4 @@
 import {
-
   required as _required,
   email as _email,
   minLength as _minLength,
@@ -14,7 +13,8 @@ import { i18n } from '@/localization'
 
 const { t } = i18n.global || i18n
 
-const messagePath = ({ $validator }: MessageProps) => `validations.field-error_${$validator}`
+const messagePath = ({ $validator }: MessageProps) =>
+  `validations.field-error_${$validator}`
 
 const withI18nMessage = createI18nMessage({ t, messagePath })